Create README.md
This project is a simple project built in C language.

This project has following features –

It displays a nicely formatted calendar of every day of every month.
The calendar application presented here is a very simple console application developed using C programming language.
It is compiled in Eclipse using eclipse java compiler.
The Calenadr Project is developed using C programming language. This Calendar in c helps the user to view the calendar of any desired year in a neat and convenient manner.